Move out scope extensions from mapper extensions
